Move controlled CU fix into CUGate
